Add DDR 573 module.

+++ Source/RhythmCodex.Lib/Ddr/Models/Ddr573DirectoryEntry.cs
@@ -3,12 +3,12 @@ namespace RhythmCodex.Ddr.Models
     public class Ddr573DirectoryEntry
     {
         public int Id { get; set; }
-        public short Offset { get; set; }
-        public short Module { get; set; }
-        public byte CompressionType { get; set; }
-        public byte Reserved0 { get; set; }
-        public byte Reserved1 { get; set; }
-        public byte Reserved2 { get; set; }
+        public int Offset { get; set; }
+        public int Module { get; set; }
+        public int CompressionType { get; set; }
+        public int Reserved0 { get; set; }
+        public int Reserved1 { get; set; }
+        public int Reserved2 { get; set; }
         public int Length { get; set; }
     }
 }
