+++ Source/RhythmCodex.Lib/Ddr/Models/Ddr573File.cs
@@ -5,9 +5,9 @@ namespace RhythmCodex.Ddr.Models
         public int Id { get; set; }
         public int Offset { get; set; }
         public int Module { get; set; }
-        public byte Reserved0 { get; set; }
-        public byte Reserved1 { get; set; }
-        public byte Reserved2 { get; set; }
+        public int Reserved0 { get; set; }
+        public int Reserved1 { get; set; }
+        public int Reserved2 { get; set; }
         public byte[] Data { get; set; }
     }
 }
